refactor(spec): extract validate-map-forms and tier-2 temporal tests
Move channel/DataRef form classification out of the TypeBox error mapper
into validate-map-forms.ts, and split temporal decision-reuse tier-2 tests
from the general data-aware suite.
